langserve_api/pyproject.toml
[tool.poetry]
name = "cohere_ner_extraction_api"
version = "0.1.0"
description = ""
authors = ["Rishov Chatterjee <rishov.chatterjee@pathrise.com>", "Ali Ahmed <hxhhisoka0@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain_openai = "^0.0.8"
pydantic = "<2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^1.4.51"
pandas = "<2"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
